improve performance

+++ src/org/honghe/BranchNode.java
@@ -14,7 +14,7 @@ public class BranchNode extends Node{
 	 */
 	public BranchNode(char c) {
 		super.setC(c);
-		super.setType(1);
+		super.setType((byte) 1);
 		super.setCount(0);
 		super.setChildren(new HashMap<String, Node>());
 	}
