+++ src/org/honghe/LeafNode.java
@@ -10,6 +10,6 @@ public class LeafNode extends Node{
 	public LeafNode() {
 		super.setC('$');
 		super.setCount(0);
-		super.setType(1);
+		super.setType((byte) 1);
 	}
 }
